Docker/app/requirements.txt
Flask
mysql-connector
pyVCF